navigate now works with room entity set - thanks @LE-tarantino!
