+++ src/hass/data/area_registry.ts
@@ -6,6 +6,6 @@ import type { RegistryEntry } from './registry';
 
 export interface AreaRegistryEntry extends RegistryEntry {
   area_id: string;
-  icon: string | null;
+  icon: string | undefined;
   name: string;
 }
