+++ test/delegates/index.spec.ts
@@ -1,12 +1,9 @@
 import actionHandlerDelegate from './action-handler-delegate.spec';
 import retrieversDelegate from './retrievers/index.spec';
-import cardEntitiesSpec from './utils/card-entities.spec';
+import utilsSpec from './utils/index.spec';
 
 describe('delegates', () => {
   retrieversDelegate();
   actionHandlerDelegate();
-
-  describe('utils', () => {
-    cardEntitiesSpec();
-  });
+  utilsSpec();
 });
