+++ test/delegates/utils/index.spec.ts
@@ -1,0 +1,9 @@
+import cardEntitiesSpec from './card-entities.spec';
+import roomEntitySpec from './room-entity.spec';
+
+export default () => {
+  describe('utils', () => {
+    cardEntitiesSpec();
+    roomEntitySpec();
+  });
+};
